move around tests to more closely match source tree write a few missed tests
